support tool call approvals and hitl stall replan
